create note list activity and fragment.
