+++ app/src/main/java/com/classified/ogomb/notetracker/NoteActivity.java
@@ -6,23 +6,10 @@ import android.os.Bundle;
 import android.support.v4.app.FragmentManager;
 
 
-public class NoteActivity extends FragmentActivity {
+public class NoteActivity extends SingleFragmentActivity {
 
     @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_fragment);
-
-
-        FragmentManager manager = getSupportFragmentManager();
-        Fragment fragment = manager.findFragmentById(R.id.fragment_container);
-
-        if (fragment == null){
-            fragment = new NoteFragment();
-            manager.beginTransaction()
-                    .add(R.id.fragment_container, fragment)
-                    .commit();
-        }
-
+    public Fragment createFragment() {
+        return new NoteFragment();
     }
 }
