+++ app/src/main/java/com/classified/ogomb/notetracker/NoteListFragment.java
@@ -1,0 +1,10 @@
+package com.classified.ogomb.notetracker;
+
+import android.support.v4.app.Fragment;
+
+/**
+ * Created by lewismbogo on 09/04/2018.
+ */
+
+public class NoteListFragment extends Fragment {
+}
