+++ app/src/main/java/com/classified/ogomb/notetracker/SingleFragmentActivity.java
@@ -1,0 +1,33 @@
+package com.classified.ogomb.notetracker;
+
+import android.os.Bundle;
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentActivity;
+import android.support.v4.app.FragmentManager;
+
+/**
+ * Created by lewismbogo on 09/04/2018.
+ */
+
+public abstract class SingleFragmentActivity extends FragmentActivity {
+
+    public abstract Fragment createFragment();
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_fragment);
+
+
+        FragmentManager manager = getSupportFragmentManager();
+        Fragment fragment = manager.findFragmentById(R.id.fragment_container);
+
+        if (fragment == null){
+            fragment = createFragment();
+            manager.beginTransaction()
+                    .add(R.id.fragment_container, fragment)
+                    .commit();
+        }
+
+    }
+}
